+++ src/main/java/user/dao/UserDAO.java
@@ -1,5 +1,8 @@
 package user.dao;
 
+import java.util.HashMap;
+import java.util.Map;
+
 import org.apache.ibatis.annotations.Mapper;
 
 import user.bean.UserDTO;
@@ -10,5 +13,9 @@ public interface UserDAO {
 	public String checkUserId(String userId);
 
 	public void joinUser(UserDTO userDTO);
+
+	public UserDTO loginSH(Map<String, String> map);
+
+	public String checkNaverLoginId(String userId);
 	
 }
