+++ src/main/resources/mapper/userMapper.xml
@@ -7,7 +7,15 @@
 	
 	</insert>
 	
+	<select id="loginSH" parameterType="map" resultType="userDTO">
+		select * from user where userid = #{userid} and pwd = #{pwd} and logintype = 'SH'
+	</select>
+	
+	<select id="checkNaverLoginId" resultType="userDTO">
+		select * from user where userid = #{userId} and logintype = 'NAVER'
+	</select>
+	
 	<select id="checkUserId">
-		select userid from user where userid = #{userId}
+		select count(*) from user where userid = #{userId}
 	</select>
 </mapper>
