+++ src/main/webapp/WEB-INF/views/index.jsp
@@ -13,13 +13,7 @@
 <link rel="stylesheet" href="/SpringHotel/resources/css/footer.css">
 </head>
 <body>
-	<c:if test="${sessionScope.userName != null }">
-		<jsp:include page="./common/header.jsp" />
-	</c:if>
-	
-	<c:if test="${sessionScope.adminId != null }">
-		<jsp:include page="./admin/header.jsp" />
-	</c:if>
+	<jsp:include page="./common/header.jsp" />
 	
 	<div class="container" data-view="index">
