+++ src/main/webapp/resources/css/joinCSS.css
@@ -4,17 +4,14 @@
 @import url(http://fonts.googleapis.com/earlyaccess/hanna.css);
 @import url('/SpringHotel/resources/css/footer.css');
 
-/* joinForm.jsp . joinProc.jsp*/
-.joinlist{
-	width:40%;
-	height:8%;
-	text-align:center;
-	border: 1px solid #e3e3e3;
-	background-color:  #e3e3e3;
-	font-size:larger;
-} 
+html,body{
+	margin:0;
+	width:100%;
+	height:100%;
+	font-family: 'Noto Sans KR', sans-serif;
+}
 
-#btn{
+.btn{
 	margin-top:30px;
 	margin-bottom:30px;
 	background-color: #382f24;
@@ -28,212 +25,29 @@
 
 /* join.jsp에만 적용 */
 #joinTable {
-	margin-top:50px;
-	margin-bottom:90px;
-	height:700px;
-	width:40%;
+	margin:5% auto;
+	height: 700px;
+	width: 50%;
 	
 	background-color: white;
 	border-radius: 1.5em;
 	border: 5px solid #f8f9fa;
 }
 
-#comment{
-	height:15%;
-	padding-top:30px;
+#joinTable th {
+	width: 100px;
+	padding-left: 20px;
 }
 
-input{
+.input{
 	 height:25px;
 	 border-right: 0px;
      border-left: 0px;
      border-top: 0px;
      border-bottom: #9f8a7b 1px solid;
-     font-size:larger;
+     font-size:medium;
 }
 
-
-#list1{
-	border:  1px solid #e3e3e3;
-	background-color:#382f24;
-	color:#ffffff;
-	font-weight:900;
-}
-
-#list2{
-	border: 1px solid #e3e3e3;
-	background-color:#ffffff;
-	color:#382f24;
-}
-
-/*joinProc.jsp에만 적용 */
-.joinFinish{
-	margin-top:80px;
-	margin-bottom:90px;
-	height:30%;
-	width:40%;
-	background:#f1ebd5;
-	text-align:center;
-	border-radius: 1em;
-}
-
-#list3{
-	border: 1px solid #e3e3e3;
-	background-color:#ffffff;
-	color:#382f24;
-}
-
-#list4{
-	border:  1px solid #e3e3e3;
-	background-color:#382f24;
-	color:#ffffff;
-	font-weight:900;
-}
-
-#btnProc{
-	margin-top:50px;
-	background-color: #382f24;
-    width:130px;
-    height:70px;
-    border: none;
-    border-radius: 1em;
-    font-size:large;
-    color:#f1ebd5;
-}
-
-/*idCheck.jsp에만 적용 */
-#btnID{
-	margin-top:30px;
-	margin-bottom:30px;
-	background-color: #382f24;
-    width:100px;
-    height:40px;
-    border: none;
-    border-radius: 1em;
-    font-size:small;
-    color:#f1ebd5;
-}
-
-#idCheck{
-	 height:25px;
-	 margin-left:40px;
-     border-bottom: #9f8a7b 1px solid;
-     font-size:small;
-
-}
-
-@charset "UTF-8";
-
-*{
-	margin: 0;
-	padding: 0;
-}
-
-#main {
-	width: 500px;
-	height: 700px;
-	margin: auto;
-	border: 1px solid;
-	border-color: #0d6efd;
-	
-	margin-top: 100px; 
-}
-
-#main h1 {
-	margin-bottom: 50px;
-	margin-top: 10px;
-	font-size: 30px;
-	padding-top: 50px;
-	padding-left: 30px;
-}
-
-
-html,body{
-	margin:0;
-	width:100%;
-	height:100%;
-	font-family: 'Noto Sans KR', sans-serif;
-}
-
-/*login 폴더의 footer 태그에만 적용*/
-.wrapper{
-	width:80%;   /* 너비 */
-	height:70%;
-	margin:0 auto;  /* 화면 중앙에 배치 */
-	position:relative;  /* 슬라이드 위치를 잡기 위해 */
-	text-align:center;  /* 슬라이드를 화면 중앙에 배치 */
-}
-
-#container {   
-	width:100%;   /* 너비 */
-	height:100%;
-	position:absolute;
-}
-
-#container > button {
-	position:absolute;   /* 배치 방법 : absolute */
-	height:100%;  /* 버튼 높이를 슬라이드 영역에 꽉 채움 */
-	top:0;  /* 위로 바짝 붙여서 배치 */
-	border:none;  /* 테두리 없음 */
-	padding:20px; /* 패딩 20px */
-	background-color:transparent;  /* 투명 배경 */
-	color:#ffffff;  /* 화살표 색 */
-	font-weight:1000;  /* 진하기 */
-	font-size:30px;  /* 화살표 크기 */
-	opacity:1;  /* 불투명도 */      
-}
-
-#prev {
-	left:0;  /* 왼쪽으로 바짝 붙여서 배치 */
-}
-
-#next {
-	right:0;  /* 오른쪽 바짝 붙여서 배치 */
-}
-
-.roominfo	{
-	margin:6% auto;
-	width:83%;
-	font-size:xx-large;
-	text-align:center;
-}
-
-.main{
-	width:83%;   /* 너비 */
-	height:30%;
-	margin:1% 9% ;  /* 화면 중앙에 배치 */
-	position:relative;  /* 슬라이드 위치를 잡기 위해 */
-	text-align:center;  /* 슬라이드를 화면 중앙에 배치 */
-	
-}
-
-.main2{
-	width:83%;   /* 너비 */
-	height:50%;
-	margin:5px auto ;  /* 화면 중앙에 배치 */
-	position:relative;  /* 슬라이드 위치를 잡기 위해 */
-	text-align:center;  /* 슬라이드를 화면 중앙에 배치 */
-}
-
-/*join.jsp*/
-.log{
-	background:#f1ebd5;
-	margin:5% auto;
-	width:470px;
-	height:570px;
-	border-radius: 1em;
-}
-
-
-.log td {
-	width: 100px;
-}
-
-.log th {
-	width: 50px;
-	margin-left: 120px;
-	background-color: green;
-}
 
 .log #title{
 	height:110px;
@@ -241,14 +55,14 @@ html,body{
 	text-align: center;
 }
 
-.input{
+.input{   
 	 height:35px;
 	 border-right: 0px;
      border-left: 0px;
      border-top: 0px;
      border-bottom: #9f8a7b 1px solid;
      border-radius: 0.2em;
-     font-size:medium;
+     font-size: medium;
      padding-left: 7px;
 }
 
@@ -272,16 +86,11 @@ html,body{
 	background-color: #45392a;
 }
 
-#loginBtn {
-	margin-top: 12px;
-}
-
 #welcomText {
 	margin-bottom: 35px;
 	font-size:larger;
 }
 
-#SocialText {
-	margin-top: 40px;
-	margin-bottom: 0;
+#loginBtn {
+	margin-top: 12px;
 }
