+++ src/main/webapp/resources/css/loginCSS.css
@@ -31,24 +31,7 @@ html,body{
  	color: black; 
  	text-decoration: none;	
 }
-/* 모든 footer 태그에 적용*/
-footer{
-	bottom: 0;
-	width: 100%;
-}
-
 /*login 폴더의 footer 태그에만 적용*/
-/* footer input{
-	margin-top:15px;
-	margin-bottom:15px;
-	background-color: #382f24;
-	width:100px;
-	height:50px;
-	border: none;
-	font-size:medium;
-	color:#f1ebd5;
-} */
-
 .wrapper{
 	width:80%;   /* 너비 */
 	height:70%;
